added optimization based on the strides for 2D arrays
