+++ quspin/operators/_oputils/openmp.h
@@ -3,6 +3,26 @@
 
 #if defined(_OPENMP)
 #include <omp.h>
+
+
+
+template<class T>
+void inline atomic_add(T &y,const T &aa){
+    #pragma omp atomic
+    y += aa;
+}
+
+template<class T>
+void inline atomic_add(std::complex<T> &y,const std::complex<T> &aa){
+    T * y_v = reinterpret_cast<T*>(&y);
+    const T * aa_v = reinterpret_cast<const T*>(&aa);
+
+    #pragma omp atomic
+    y_v[0] += aa_v[0];
+    #pragma omp atomic
+    y_v[1] += aa_v[1];    
+}
+
 #else
 typedef int omp_int_t;
 inline omp_int_t omp_get_thread_num() { return 0;}
